Added pagination in category route
