--- src/routes/category/[slug]/+page.server.ts
@@ -1,10 +1,0 @@
-import { getArticlesByCategory } from '$lib/data/category/api.js';
-
-export async function load({ params }) {
-	const { slug: category } = params;
-	if (!category) {
-		return { status: 404 };
-	}
-	const articles = await getArticlesByCategory(category);
-	return articles;
-}
